+++ dwarfeng-dct-core/src/main/java/com/dwarfeng/dct/node/configuration/DataCodingConfigDefinitionParser.java
@@ -33,7 +33,9 @@ public class DataCodingConfigDefinitionParser implements BeanDefinitionParser {
 
         BeanDefinitionParserUtil.makeSureBeanNameNotDuplicated(parserContext, configName);
 
-        RootBeanDefinition dataCodingConfigBuilderBeanDefinition = new RootBeanDefinition(DataCodingConfig.Builder.class);
+        RootBeanDefinition dataCodingConfigBuilderBeanDefinition = new RootBeanDefinition(
+                DataCodingConfig.Builder.class
+        );
         dataCodingConfigBuilderBeanDefinition.getPropertyValues().add(
                 "flatDataCodec", new RuntimeBeanReference(flatDataCodecRef)
         );