+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 ### 功能构建
 
+- `dwarfeng-dct-core` 子模块类优化注释、文档注释格式、代码换行格式。
+  - com.dwarfeng.dct.node.configuration.DataCodingConfigDefinitionParser。
+  - com.dwarfeng.dct.node.configuration.DataCodingQosDefinitionParser。
+  - com.dwarfeng.dct.node.configuration.ValueCodingQosDefinitionParser。
+  - com.dwarfeng.dct.node.configuration.ServiceExceptionMapperConfiguration。
+  - com.dwarfeng.dct.sdk.util.ServiceExceptionHelperTest。
+
 - 优化文件格式。
   - 优化 `application-context-*.xml` 文件的格式。
   - 优化 `pom.xml` 文件的格式。